Keep FolderListGateway under the 128KiB file-size ratchet
Move the pocketshell JSON enumerator and row union into a sibling file
so the gateway is no longer a new oversized source file.
